backend/app/integrations/token_sale_manager.py
from typing import Optional, Dict
import os
from solana.rpc.async_api import AsyncClient
from solders.keypair import Keypair
from solders.pubkey import Pubkey
from solders.system_program import create_account, CreateAccountParams
from solders.transaction import Transaction
from spl.token.constants import TOKEN_PROGRAM_ID
from spl.token.client import Token
from spl.token.instructions import (
    initialize_mint, create_associated_token_account,
    mint_to, get_associated_token_address,
    approve, revoke, set_authority
)
import base58
import asyncio
from .solana_manager import SolanaTokenManager
from dataclasses import dataclass

@dataclass
class TokenSaleConfig:
    price_usd: float  # Price per token in USD
    target_raise: float  # Total amount to raise in USD
    initial_supply: int  # Initial token supply
    decimals: int = 6  # Number of decimal places for the token

class TokenSaleManager:
    def __init__(self, solana_manager: SolanaTokenManager):
        self.solana_manager = solana_manager
        self.usdc_mint = Pubkey.from_string(os.getenv("USDC_MINT_ADDRESS", "EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v"))  # Mainnet USDC
        
    async def create_token_sale(self, name: str, symbol: str, config: TokenSaleConfig, creator_wallet: str) -> Dict:
        """Create a new token with fixed-price sale configuration"""
        # Create the token using SolanaTokenManager
        token_result = await self.solana_manager.create_token(
            name=name,
            symbol=symbol,
            initial_supply=config.initial_supply,
            creator_wallet=creator_wallet,
            features={"burnable": False, "mintable": False},  # Lock token features for Raydium compatibility
            decimals=config.decimals
        )
        
        # Store sale configuration in database (to be implemented)
        sale_config = {
            "token_address": token_result["token_address"],
            "price_usd": config.price_usd,
            "target_raise": config.target_raise,
            "initial_supply": config.initial_supply,
            "decimals": config.decimals,
            "creator_wallet": creator_wallet,
            "status": "pending"
        }
        
        return {
            **token_result,
            "sale_config": sale_config
        }
    
    async def setup_raydium_pool(self, token_address: str, creator_wallet: str) -> Dict:
        """Set up a Raydium liquidity pool for the token"""
        # This is a placeholder for Raydium integration
        # In production, you would:
        # 1. Revoke mint authority
        # 2. Create Raydium pool
        # 3. Add initial liquidity
        
        try:
            # Revoke mint authority to make token immutable
            token = Token(
                self.solana_manager.client,
                Pubkey.from_string(token_address),
                TOKEN_PROGRAM_ID,
                Keypair.from_secret_key(base58.b58decode(os.getenv("SOLANA_PAYER_KEY")))
            )
            
            await token.set_authority(
                token_address,
                None,  # New authority is None (revoked)
                "MintTokens",
                Pubkey.from_string(creator_wallet),
                [Keypair.from_secret_key(base58.b58decode(os.getenv("SOLANA_PAYER_KEY")))]
            )
            
            # TODO: Implement actual Raydium pool creation
            # For now, return mock response
            return {
                "status": "success",
                "token_address": token_address,
                "pool_address": f"mock_pool_{token_address[-6:]}",
                "message": "Raydium pool setup completed (mock)"
            }
            
        except Exception as e:
            return {
                "status": "error",
                "message": f"Failed to setup Raydium pool: {str(e)}"
            }
    
    async def process_token_purchase(self, token_address: str, buyer_wallet: str, usdc_amount: float) -> Dict:
        """Process a token purchase at the fixed price"""
        # This is a placeholder for the token purchase logic
        # In production, you would:
        # 1. Verify USDC balance
        # 2. Execute token transfer
        # 3. Update sale status
        
        try:
            # Mock successful purchase
            return {
                "status": "success",
                "token_address": token_address,
                "buyer_wallet": buyer_wallet,
                "usdc_amount": usdc_amount,
                "transaction_signature": f"mock_tx_{token_address[-6:]}_{usdc_amount}",
                "message": "Token purchase completed (mock)"
            }
            
        except Exception as e:
            return {
                "status": "error",
                "message": f"Failed to process token purchase: {str(e)}"
            } 